add qwen vit
